tweak(utils): split `filesystem` into single-responsibility modules (#6913)
