+++ packages/utils/src/index.js
@@ -9,7 +9,14 @@ export { Multilock } from './Multilock';
 export * from './period';
 export * from './testUtilities';
 export * from './featureFlags';
-export * from './filesystem';
+export { getDeDuplicatedFileName } from './getDeDuplicatedFileName';
+export { getNestedFiles } from './getNestedFiles';
+export { getUniqueFileNameParts } from './getUniqueFileNameParts';
+export { readJsonFile } from './readJsonFile';
+export { sanitizeWorksheetName } from './sanitizeWorksheetName';
+export { toFilename } from './toFilename';
+export { writeJsonFile } from './writeJsonFile';
+export { writeStreamToFile } from './writeStreamToFile';
 export { filterEntities } from './filterEntities';
 export * from './geoJson';
 export { getCountryNameFromCode } from './getCountryNameFromCode';
