+++ packages/utils/src/writeJsonFile.js
@@ -1,0 +1,4 @@
+import fs from 'node:fs';
+
+export const writeJsonFile = (filePath, json) =>
+  fs.writeFileSync(filePath, `${JSON.stringify(json, null, 2)}\n`);
